+++ app/controllers/application_controller.rb
@@ -11,8 +11,7 @@ class ApplicationController < ActionController::Base
 
   def configure_permitted_parameters
     devise_parameter_sanitizer.permit(
-      :sign_up,
-      keys: [:first_name, :last_name, :email, :password, :password_confirmation]
+      :sign_up, keys: [:username, :full_name, :email, :password]
     )
   end