+++ app/helpers/email_helper.rb
@@ -1,94 +1,98 @@
 # frozen_string_literal: true
 
-# Helper functions for application emails
+# Helper functions for application emails.
 module EmailHelper
-  def success_color
-    '#df9cca'
+  def emphasis_color
+    "#404040" # $p200-black
   end
 
-  def warning_color
-    '#cca781'
-  end
-
-  def danger_color
-    '#bf9a74'
+  def muted_color
+    "#989ea6"
   end
 
   def text_color
-    '#3a3b3c'
+    "#3a3b3c"
   end
 
   def background_color
-    '#b295db'
+    "#512da8" # $p700-deep-purple
+  end
+
+  def banner_color
+    emphasis_color
   end
 
   def link_color
-    '#52357b'
+    "#673ab7" # $p500-deep-purple
   end
 
   def center_style
     hash_to_css_string(
-      font_size: '17px',
-      line_height: '24px',
-      margin: '0 0 16px',
-      text_align: 'center'
+      font_size: "17px",
+      line_height: "24px",
+      margin: "0 0 16px",
+      text_align: "center"
     )
   end
 
   def link_style
     hash_to_css_string(
       color: link_color,
-      font_weight: 'bold',
-      text_decoration: 'none',
-      word_break: 'break-word'
+      font_weight: "bold",
+      word_break: "break-word"
     )
   end
 
   def p_style
     hash_to_css_string(
-      font_size: '17px',
-      line_height: '24px',
-      margin: '0 0 16px'
+      font_size: "17px",
+      line_height: "24px",
+      margin: "0 0 16px"
+    )
+  end
+
+  def image_style
+    hash_to_css_string(
+      text_align: "center",
+      max_width: "100%"
     )
   end
 
   def blockquote_style
     hash_to_css_string(
-      font_size: '17px',
-      font_style: 'italic',
-      line_height: '24px',
-      margin: '0 0 16px'
+      font_size: "17px",
+      font_style: "italic",
+      line_height: "24px",
+      margin: "0 0 16px"
     )
   end
 
   def default_style
     hash_to_css_string(
-      font_weight: 'bold',
-      word_break: 'break-word'
+      font_weight: "bold",
+      word_break: "break-word"
     )
   end
 
-  def success_style
+  def emphasis_style
     hash_to_css_string(
-      color: success_color,
-      font_weight: 'bold',
-      word_break: 'break-word'
+      color: emphasis_color,
+      font_weight: "bold",
+      word_break: "break-word"
     )
   end
 
-  def warning_style
+  def regular_style
     hash_to_css_string(
-      color: warning_color,
-      font_weight: 'bold',
-      word_break: 'break-word'
+      color: text_color,
+      word_break: "break-word"
     )
   end
 
-  def danger_style
+  def muted_style
     hash_to_css_string(
-      color: danger_color,
-      font_weight: 'bold',
-      word_break: 'break-word'
+      color: muted_color,
+      word_break: "break-word"
     )
   end
 
@@ -98,6 +102,6 @@ module EmailHelper
     array = hash.collect do |key, value|
       "#{key.to_s.dasherize}:#{value}"
     end
-    array.join(';')
+    array.join(";")
   end
 end