+++ test/mailers/devise_mailer_test.rb
@@ -1,15 +1,14 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Tests links in password reset email
 class DeviseMailerTest < ActionMailer::TestCase
-  test 'devise reset password email' do
+  test "devise reset password email" do
     user = users(:viewer)
-    email = Devise::Mailer.reset_password_instructions(user, 'faketoken').deliver_now
-    assert !ActionMailer::Base.deliveries.empty?
-    assert_equal [user.email], email.to
-    assert_equal 'Reset password instructions', email.subject
-    assert_match(%r{#{ENV['website_url']}/password/edit\?reset_password_token=faketoken}, email.encoded)
+    mail = Devise::Mailer.reset_password_instructions(user, "faketoken")
+    assert_equal "Reset password for your PATS Trial account", mail.subject
+    assert_equal [user.email], mail.to
+    assert_match(%r{#{ENV["website_url"]}/password/edit\?reset_password_token=faketoken}, mail.body.encoded)
   end
 end